Get a count of detections per recorder grouped by given parameters.

Wrapper around the 'aggregations_project_recorders_count_detections' endpoint to retrieve a count of detections for each recorder and grouped by the specified query parameters.

Usage

get_recorders_count_detections(project_name, ...)

Arguments

project_name

Name of the project to get a count on detetcions class and recorder

...

query paramaters. See http://192.168.10.30:8001/api/docs/#tag/v0.2/operation/v0.2_aggregations_project_recorder_recordings_count_detections_retrieve

Value

A summary containing species counts per recorder within a project that match the specified query parameters: http://192.168.10.30:8001/api/docs/#tag/v0.2/operation/v0.2_aggregations_project_recorder_recordings_count_detections_retrieve.

Examples

# Retrieve a count pre species and recorders. By default, the count is returned for today
if (FALSE) { # \dontrun{
get_recorders_count_detections(project_name = "pam_in_chemnitz")
} # }
# You can specify a start_date and end_date to get a count for specific time interval.
# Limit 'countable' detections by setting a confidence threshold
if (FALSE) { # \dontrun{
get_recorders_count_detections(project_name = "pam_in_chemnitz", start_datetime = "2023-01-01", end_datetime = "2023-12-31", min_confidence = 0.85)
} # }